requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.4
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
pymongo==3.11.4
pytz==2021.1
requests==2.25.1
schedule==1.1.0
soupsieve==2.2.1
sqlparse==0.2.4
urllib3==1.26.4
waitress==2.0.0
whitenoise==5.2.0
